refactor(tests): refactorizar auth.service.spec.ts - dividir en archivos especializados
- Dividir auth.service.spec.ts (573 líneas) en 2 archivos especializados
- Crear auth.service.authentication.spec.ts (205 líneas): register, login, validateUser
- Crear auth.service.tokens.spec.ts (239 líneas): refreshToken, token generation, sanitization
- Extraer helpers comunes a helpers/auth.test-helpers.ts (167 líneas)
- Implementar factories pattern: createMockUser, createRegisterDto, createLoginDto
- Implementar setup helper: setupAuthTestModule() para reutilización
- Organizar tests con describe anidados para mejor legibilidad

Archivos:
- auth.service.authentication.spec.ts: 17 tests (autenticación core)
- auth.service.tokens.spec.ts: 12 tests (tokens y seguridad)
- helpers/auth.test-helpers.ts: factories y setup utilities

Tests: 85 passing (29 nuevos del archivo dividido)
Coverage módulo auth: 87.02% statements, 93.75% functions
Promedio: ~204 líneas/archivo (target: 150-250)

Documentación:
- REFACTOR_TEST_INVENTORY.md: inventario completo de tests
- docs/refactor/auth-module-analysis.md: análisis del módulo Auth
